feat: refine write nav  #161
- replaced nav items array with writeConfig
- updated write's nav to write-nav
- updated items
- removed settings because it's redundant
- hid analytics from menu for now until can iterate
- enabled Tools in main-nav

Closes #161
