+++ app/config/admin.ts
@@ -19,7 +19,6 @@ export const adminConfig: AdminConfig = {
     {
       title: 'Tools',
       href: '/tools',
-      disabled: true,
     },
     {
       title: 'Insights',
