+++ app/config/dashboard.ts
@@ -19,7 +19,6 @@ export const dashboardConfig: DashboardConfig = {
     {
       title: 'Tools',
       href: '/tools',
-      disabled: true,
     },
     {
       title: 'Insights',
