+++ app/config/marketing.ts
@@ -24,7 +24,6 @@ export const marketingConfig: MarketingConfig = {
     {
       title: 'Tools',
       href: '/tools',
-      disabled: true,
     },
     {
       title: 'Insights',
