+++ app/config/playbook.ts
@@ -19,7 +19,6 @@ export const playbookConfig: PlaybookConfig = {
     {
       title: 'Tools',
       href: '/tools',
-      disabled: true,
     },
     {
       title: 'Insights',
