+++ app/config/teacher.ts
@@ -19,7 +19,6 @@ export const teacherConfig: TeacherConfig = {
     {
       title: 'Tools',
       href: '/tools',
-      disabled: true,
     },
     {
       title: 'Insights',
